feat: 추가 수정 및 디렉터리 수정
