+++ backend/src/main/java/SANTA/backend/core/auth/dto/LoginRequestDTO.java
@@ -1,0 +1,11 @@
+package SANTA.backend.core.auth.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class LoginRequestDTO {
+    private String username;
+    private String password;
+}
