+++ backend/src/main/java/SANTA/backend/core/auth/dto/LoginResponseDTO.java
@@ -1,0 +1,13 @@
+package SANTA.backend.core.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class LoginResponseDTO {
+    private String token;
+}
