+++ backend/src/main/java/SANTA/backend/core/auth/jwt/JWTProperties.java
@@ -1,0 +1,19 @@
+package SANTA.backend.core.auth.jwt;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Component
+@ConfigurationProperties(prefix = "spring.jwt")
+public class JWTProperties {
+
+    private String secret;
+
+    public String getSecret() {
+        return secret;
+    }
+
+    public void setSecret(String secret) {
+        this.secret = secret;
+    }
+}
