+++ backend/src/main/java/SANTA/backend/core/auth/jwt/LoginFilter.java
@@ -1,0 +1,52 @@
+package SANTA.backend.core.auth.jwt;
+
+import SANTA.backend.core.auth.dto.LoginRequestDTO;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.security.authentication.AuthenticationServiceException;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+import java.io.IOException;
+
+public class LoginFilter extends UsernamePasswordAuthenticationFilter {
+    @Override
+    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws
+            AuthenticationException {
+        if (!request.getMethod().equals("POST")) {
+            throw new AuthenticationServiceException(
+                    "Authentication method not supported: " + request.getMethod());
+        }
+
+        LoginRequestDTO loginRequest;
+        try {
+            loginRequest = new ObjectMapper().readValue(request.getInputStream(), LoginRequestDTO.class);
+        } catch (IOException e) {
+            throw new AuthenticationServiceException("로그인 실패");
+        }
+
+        String username = loginRequest.getUsername();
+        String password = loginRequest.getPassword();
+
+        if (username == null) {
+            username = "";
+        }
+
+        if (password == null) {
+            password = "";
+        }
+
+        username = username.trim();
+
+        UsernamePasswordAuthenticationToken authRequest = new UsernamePasswordAuthenticationToken(
+                username, password);
+
+        // Allow subclasses to set the "details" property
+        setDetails(request, authRequest);
+
+        return this.getAuthenticationManager().authenticate(authRequest);
+    }
+}
