+++ backend/src/main/java/SANTA/backend/core/auth/jwt/LoginSuccessHandler.java
@@ -1,0 +1,30 @@
+package SANTA.backend.core.auth.jwt;
+
+
+import SANTA.backend.core.auth.dto.LoginResponseDTO;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+import org.springframework.stereotype.Component;
+
+@RequiredArgsConstructor
+@Component
+public class LoginSuccessHandler implements AuthenticationSuccessHandler {
+    private final JWTUtil tokenProvider;
+
+    @Override
+    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response,
+                                        Authentication authentication) throws java.io.IOException, ServletException {
+        response.setStatus(HttpStatus.OK.value());
+        response.setContentType("application/json");
+        response.setCharacterEncoding("UTF-8");
+        new ObjectMapper().writeValue(response.getWriter(),
+                new LoginResponseDTO(tokenProvider.createJwt(authentication)));
+    }
+}
+
