+++ backend/src/main/java/SANTA/backend/core/auth/oauth/GoogleUserInfo.java
@@ -1,0 +1,35 @@
+package SANTA.backend.core.auth.oauth;
+
+import java.util.Map;
+
+public class GoogleUserInfo extends OAuth2UserInfo {
+
+    public GoogleUserInfo(Map<String, Object> attributes) {
+        super(attributes);
+    }
+
+    @Override
+    public String getId() {
+        return (String) attributes.get("sub");
+    }
+
+    @Override
+    public String getUsername() {
+        return (String) attributes.get("email");
+    }
+
+    @Override
+    public String getName() {
+        return (String) attributes.get("name");
+    }
+
+    @Override
+    public String getEmail() {
+        return (String) attributes.get("email");
+    }
+
+    @Override
+    public String getProvider() {
+        return "Google";
+    }
+}
