--- backend/src/main/java/SANTA/backend/core/controller/AdminController.java
@@ -1,15 +1,0 @@
-package SANTA.backend.core.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@Controller
-@ResponseBody
-public class AdminController {
-    @GetMapping("/admin")
-    public String adminP(){
-
-        return "admin Controller";
-    }
-}
