--- backend/src/main/java/SANTA/backend/core/controller/MainController.java
@@ -1,32 +1,0 @@
-package SANTA.backend.core.controller;
-
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-import java.util.Collection;
-import java.util.Iterator;
-
-@Controller
-@ResponseBody
-public class MainController {
-
-    @GetMapping("/abc")
-    public String mainP(){
-
-        //username 가져오기
-        String username=  SecurityContextHolder.getContext().getAuthentication().getName();
-        //role값 가져오기
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-        Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
-        Iterator<? extends GrantedAuthority> iter = authorities.iterator();
-        GrantedAuthority auth = iter.next();
-        String role = auth.getAuthority();
-
-        return "abc";
-    }
-}
