--- backend/src/main/java/SANTA/backend/core/domain/Role.java
@@ -1,5 +1,0 @@
-package SANTA.backend.core.domain;
-
-public enum Role {
-    ROLE_USER, ROLE_ADMIN
-}
