--- backend/src/main/java/SANTA/backend/core/dto/JoinDTO.java
@@ -1,15 +1,0 @@
-package SANTA.backend.core.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-public class JoinDTO {
-
-    private String username;
-    private String password;
-    private String nickname;
-    private int age;
-    // username, nickname, password, age UserEntity확인
-}
