--- backend/src/main/java/SANTA/backend/core/dto/LoginRequestDTO.java
@@ -1,11 +1,0 @@
-package SANTA.backend.core.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class LoginRequestDTO {
-    private String username;
-    private String password;
-}
