--- backend/src/main/java/SANTA/backend/core/dto/LoginResponseDTO.java
@@ -1,13 +1,0 @@
-package SANTA.backend.core.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class LoginResponseDTO {
-    private String token;
-}
