--- backend/src/main/java/SANTA/backend/core/service/UserService.java
@@ -1,9 +1,0 @@
-package SANTA.backend.core.service;
-
-
-import SANTA.backend.core.domain.User;
-
-public interface UserService {
-    public User findById(String userId);
-    public User register(User user);
-}
