+++ backend/src/main/java/SANTA/backend/core/user/domain/Role.java
@@ -1,0 +1,5 @@
+package SANTA.backend.core.user.domain;
+
+public enum Role {
+    ROLE_USER, ROLE_ADMIN
+}
