--- backend/src/main/java/SANTA/backend/global/jwt/JWTProperties.java
@@ -1,19 +1,0 @@
-package SANTA.backend.global.jwt;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Component
-@ConfigurationProperties(prefix = "spring.jwt")
-public class JWTProperties {
-
-    private String secret;
-
-    public String getSecret() {
-        return secret;
-    }
-
-    public void setSecret(String secret) {
-        this.secret = secret;
-    }
-}
