--- backend/src/main/java/SANTA/backend/global/security/userinfo/CustomUserDetails.java
@@ -1,93 +1,0 @@
-package SANTA.backend.global.security.userinfo;
-
-import SANTA.backend.core.domain.User;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Map;
-
-public class CustomUserDetails implements UserDetails, OAuth2User {
-
-    private final User user;
-
-    private Map<String, Object>attributes;
-
-    private CustomUserDetails(User user) {
-        this.user = user;
-    }
-
-    public static CustomUserDetails create(User user) {
-        return new CustomUserDetails(user);
-    }
-
-    public static CustomUserDetails create(User user, Map<String, Object> attributes) {
-        CustomUserDetails userPrincipal = CustomUserDetails.create(user);
-        userPrincipal.setAttributes(attributes);
-        return userPrincipal;
-    }
-
-    public void setAttributes(Map<String, Object> attributes) {
-        this.attributes = attributes;
-    }
-
-
-    @Override
-    public Map<String, Object> getAttributes() {
-        return this.attributes;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        Collection<GrantedAuthority> collection=new ArrayList<>();
-        collection.add(new GrantedAuthority() {
-            @Override
-            public String getAuthority() {
-                return user.getRole().name();
-            }
-        });
-        return collection;
-    }
-
-    @Override
-    public String getPassword() {
-        return user.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return user.getUsername();
-    }
-
-    public Long getUserId(){return user.getUserId();}
-
-    public String getNickname(){return user.getNickname();}
-
-    //계정에 관한 추가 설정들
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-
-    @Override
-    public String getName() {
-        return "";
-    }
-}
