--- backend/src/main/java/SANTA/backend/global/security/userinfoservice/CustomOauth2UserService.java
@@ -1,65 +1,0 @@
-package SANTA.backend.global.security.userinfoservice;
-
-import SANTA.backend.core.domain.User;
-import SANTA.backend.core.repository.UserRepository;
-import SANTA.backend.global.security.userinfo.CustomUserDetails;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.authentication.InternalAuthenticationServiceException;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.stereotype.Service;
-import org.springframework.util.StringUtils;
-
-import java.util.Optional;
-
-@RequiredArgsConstructor
-@Service
-@Slf4j
-public class CustomOauth2UserService extends DefaultOAuth2UserService {
-
-    private final UserRepository userRepository;
-
-
-    @Override
-    public OAuth2User loadUser(OAuth2UserRequest oAuth2UserRequest) {
-        OAuth2User oAuth2User = super.loadUser(oAuth2UserRequest);
-
-        try {
-            return processOAuth2User(oAuth2UserRequest, oAuth2User);
-        } catch (AuthenticationException ex) {
-            throw ex;
-        } catch (Exception ex) {
-            throw new InternalAuthenticationServiceException(ex.getMessage(), ex.getCause());
-        }
-    }
-
-    private OAuth2User processOAuth2User(OAuth2UserRequest oAuth2UserRequest, OAuth2User oAuth2User) throws org.apache.tomcat.websocket.AuthenticationException {
-        OAuth2UserInfo oAuth2UserInfo = OAuth2UserInfoFactory.getOAuth2UserInfo(
-                oAuth2UserRequest.getClientRegistration().getRegistrationId(), oAuth2User.getAttributes());
-        if (StringUtils.isEmpty(oAuth2UserInfo.getEmail())) {
-            log.info("@@@@@@@@@@@@@@@@@@@@@@username 정보 없음@@@@@@@@@@@@@@@@@@@@");
-            throw new org.apache.tomcat.websocket.AuthenticationException("username 정보 없음;");
-        }
-
-        Optional<User> userOptional = Optional.ofNullable(userRepository.findByUsername(oAuth2UserInfo.getUsername()));
-        User user;
-        if (userOptional.isPresent()) {
-            user = userOptional.get();
-        } else {
-            user = registerNewUser(oAuth2UserRequest, oAuth2UserInfo);
-        }
-
-        return CustomUserDetails.create(user, oAuth2User.getAttributes());
-    }
-
-    private User registerNewUser(OAuth2UserRequest oAuth2UserRequest, OAuth2UserInfo oAuth2UserInfo) {
-        User user = User.registerUser(oAuth2UserInfo.getUsername(), oAuth2UserInfo.getId(),oAuth2UserInfo.getName());
-        String provider = oAuth2UserRequest.getClientRegistration().getRegistrationId();
-
-        return userRepository.join(user);
-    }
-
-}
