--- backend/src/main/java/SANTA/backend/global/security/userinfoservice/GoogleUserInfo.java
@@ -1,35 +1,0 @@
-package SANTA.backend.global.security.userinfoservice;
-
-import java.util.Map;
-
-public class GoogleUserInfo extends OAuth2UserInfo {
-
-    public GoogleUserInfo(Map<String, Object> attributes) {
-        super(attributes);
-    }
-
-    @Override
-    public String getId() {
-        return (String) attributes.get("sub");
-    }
-
-    @Override
-    public String getUsername() {
-        return (String) attributes.get("email");
-    }
-
-    @Override
-    public String getName() {
-        return (String) attributes.get("name");
-    }
-
-    @Override
-    public String getEmail() {
-        return (String) attributes.get("email");
-    }
-
-    @Override
-    public String getProvider() {
-        return "Google";
-    }
-}
