--- backend/src/main/java/SANTA/backend/global/security/userinfoservice/OAuth2UserInfo.java
@@ -1,22 +1,0 @@
-package SANTA.backend.global.security.userinfoservice;
-
-import java.util.Map;
-
-public abstract class OAuth2UserInfo {
-
-    protected Map<String, Object> attributes;
-
-    public OAuth2UserInfo(Map<String, Object> attributes){
-        this.attributes = attributes;
-    }
-
-    public Map<String, Object> getAttributes(){
-        return this.attributes;
-    }
-
-    public abstract String getId();
-    public abstract String getUsername();
-    public abstract String getName();
-    public abstract String getEmail();
-    public abstract String getProvider();
-}
