--- backend/src/main/java/SANTA/backend/global/security/util/CookieUtil.java
@@ -1,60 +1,0 @@
-package SANTA.backend.global.security.util;
-
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.apache.logging.log4j.util.internal.SerializationUtil;
-import org.springframework.util.SerializationUtils;
-
-import java.util.Base64;
-import java.util.Objects;
-import java.util.Optional;
-
-public class CookieUtil {
-    private CookieUtil(){throw new IllegalStateException("CookieUtils의 인스턴스를 생성할 수 없습니다.");}
-
-    public static Optional<Cookie> getCookie(HttpServletRequest request, String name){
-        Cookie[] cookies= request.getCookies();
-
-        if(cookies != null && cookies.length > 0){
-            for (Cookie cookie : cookies) {
-                if (cookie.getName().equals(name))
-                    return Optional.of(cookie);
-            }
-        }
-        return Optional.empty();
-    }
-
-    public static void addCookie(HttpServletResponse response, String name, String value, int maxAge){
-        Cookie cookie = new Cookie(name, value);
-        cookie.setPath("/");
-        cookie.setSecure(true);
-        cookie.setHttpOnly(true);
-        cookie.setMaxAge(maxAge);
-        response.addCookie(cookie);
-    }
-
-    public static void deleteCookie(HttpServletRequest request, HttpServletResponse response, String name){
-        Cookie[] cookies = request.getCookies();
-        if(cookies != null && cookies.length > 0 ){
-            for (Cookie cookie : cookies) {
-                if(cookie.getName().equals(name)){
-                    cookie.setValue("");
-                    cookie.setPath("/");
-                    cookie.setMaxAge(0);
-                    response.addCookie(cookie);
-                }
-            }
-        }
-    }
-
-    public static String serialize(Object object){
-        return Base64.getUrlEncoder()
-                .encodeToString(SerializationUtils.serialize(object));
-    }
-
-    public static <T> T deserialize(Cookie cookie, Class<T> cls){
-        return cls.cast(SerializationUtils.deserialize(
-                Base64.getUrlDecoder().decode(cookie.getValue())));
-    }
-}
